reimport utils: move into its own folder (#4245)
